refactor(signoz): split integrations/signoz/tools/__init__.py into one package per tool (#5438)
